third_party/blink/web_tests/fast/forms/month/month-input-type-expected.txt

Tests for writing and reading .type property of HTMLInputElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.type for "month" is correctly "month".
PASS input.type for "MONTH" is correctly "month".
PASS input.type for " month " is correctly "text".
PASS successfullyParsed is true

TEST COMPLETE
